backend/requirements.txt
pytest==7.1.3
Flask==2.2.2
Flask-Cors==3.0.10
requests==2.28.1
python-dotenv==0.21.0
